feat: Combine passwordInput with input component
Collapse components because this simplifies styling and positioning of
password visibility icon with dynamic labels and error messages.
Hardcoding icon position is avoided for better maintainability;
