+++ src/libs/components/components.ts
@@ -1,5 +1,4 @@
 export { App } from "./app/app.tsx";
 export { Input } from "./input/input.tsx";
 export { Button } from "./button/button.tsx";
-export { PasswordInput } from "./password-input/password-input.tsx";
 export { RouterProvider } from "./router-provider/router-provider.tsx";
