+++ src/libs/components/input/libs/components/bare-input/styles.module.css
@@ -1,0 +1,38 @@
+.input {
+  width: 100%;
+  height: 48px;
+  font-family: "Basis Grotesque Pro", Verdana, sans-serif;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 20px;
+  padding: 0 12px;
+  border: 1.2px solid var(--light-gray);
+  border-radius: 6px;
+}
+
+.input.error {
+  border: 1.2px solid var(--red);
+}
+
+.input::placeholder {
+  color: var(--hit-gray);
+}
+
+.password-input-wrapper {
+  position: relative;
+  width: 100%;
+}
+
+.password-visibility-button {
+  position: absolute;
+  right: 15px;
+  top: 50%;
+  transform: translateY(-50%);
+  background: transparent;
+  border: none;
+}
+
+.password-input::-ms-reveal,
+.password-input::-ms-clear {
+  display: none;
+}
