+++ src/libs/components/input/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { BareInput } from "./bare-input/bare-input.tsx";
