+++ src/libs/components/input/styles.module.css
@@ -5,20 +5,13 @@
   gap: 8px;
 }
 
-.input {
-  width: 100%;
-  height: 48px;
+.labelText {
   font-family: "Basis Grotesque Pro", Verdana, sans-serif;
-  font-weight: 400;
   font-size: 15px;
-  line-height: 20px;
-  padding: 0 12px;
-  border: 1.2px solid var(--light-gray);
-  border-radius: 6px;
-}
-
-.input.error {
-  border: 1.2px solid var(--red);
+  line-height: 21px;
+  font-weight: 500;
+  letter-spacing: -0.036px;
+  color: var(--black-pearl);
 }
 
 .error-message {
@@ -27,8 +20,4 @@
   font-size: 15px;
   line-height: 20px;
   color: var(--red);
-}
-
-.input::placeholder {
-  color: var(--hit-gray);
 }
