--- src/libs/components/password-input/styles.module.css
@@ -1,18 +1,0 @@
-.password-input-wrapper {
-  position: relative;
-  width: 100%;
-}
-
-.password-visibility-button {
-  position: absolute;
-  right: 15px;
-  top: 24px;
-  transform: translateY(-50%);
-  background: transparent;
-  border: none;
-}
-
-.password-input::-ms-reveal,
-.password-input::-ms-clear {
-  display: none;
-}
