+++ src/pages/auth/libs/components/login-form/styles.module.css
@@ -54,11 +54,11 @@
   background: var(--porcelain);
 }
 
-.email-wrapper {
+.email-label {
   margin-top: 30px;
 }
 
-.password-wrapper {
+.password-label {
   margin: 25px 0 15px;
 }
